Generate build tag in src/commandline-interface/main.lisp
This used to use build.id to tag builds but a change in Jenkins made
build.id identical to the build number which is not suitable for our
purposes. Explicitly generate a build tag of the previous
form (i.e. yyyy-MM-dd_HH-mm-ss) to fix the issue and guard against
future changes

* src/commandline-interface/main.lisp (configure-buildflow-job):
  generate build tag using time.time.format instead of using build.id