"Drill down" into next-values inside dictionary values in src/project/variable-evaluation.lisp
This solves the problem of using next-value in the following situation:

  object-1:
  {
    "variables": {
      "foo": {
        { "bar": { "baz": [ "fez", "@{next-value}" ] } }
      }
    }
  }

  object-2:
  {
    "variables": {
      "foo": {
        { "bar": { "baz": [ "whoop" ] } }
      }
    }
  }

where object-1 is more specific than object-2. Without this change, the
above situation resulted in the @{next-value} expression in object-1
evaluating to { "bar": { "baz": [ "whoop" ] } } which is not wrong, but
also not very useful.

The new behavior is to record the path "bar" > "baz" when evaluating the
value of "foo" in object-1. When encountering the @{next-value}
expression, it is evaluated as before, but the recorded path is used to
"drill down" into the looked-up next value to extract the nested value
at the location equivalent to the location of the @{next-value}
expression in its containing dictionary structure.

In the above example, the effective value of "foo" in object-1 is thus

  [ "fez", "whoop" ]

which is arguably more useful than the previous behavior.

* src/project/variable-evaluation.lisp (expand): record the current path
  when evaluating :alist structures and "drill down" into next values to
  obtain sub-values at structurally equivalent positions