Use generalized orchestration in src/commandline-interface/main.lisp
Apparently, Jenkins no longer supports the build-flow plugin (and the
hard-coded orchestration stuff sucked anyway. well, actually, using
the build-flow plugin for this sucked in the first place). In any
case, something new is needed and while we're at it, we can also make
it suck less.

The new, generalized generation of orchestration jobs works much like
the generation of ordinary jobs. In particular, it is completely under
the control of templates, not hard-coded into the generator.

In more detail, the generator uses a template named "orchestration"
for the current mode and generates the jobs and aspects in that
template.

Using previous enhancements, namely the extended Groovy scripting
support and the :jobs.dependencies/groovy variable, the orchestration
behavior for the toolkit mode, previously implemented as a build-flow
job, can now be implemented as a template specifying a normal job
which executes a Groovy script when built. The Groovy script serves
the same purpose as the old build-flow, but does a much better job by
parallelizing dynamically instead of statically ahead-of-time.

* src/model/schema.lisp (define-variable :disable-orchestration-jobs):
  mark as deprecated
  (define-variable :buildflow-name): likewise
  (define-variable :buildflow.parallel?): likewise
  (define-variable :buildflow.exclude?): likewise
* src/commandline-interface/main.lisp (instantiate-projects): as a
  hack, use the first and only distribution as parent for the
  instantiated projects
  (configure-buildflow-job): removed; no longer needed, not possible
  (schedule-jobs/serial): likewise
  (schedule-jobs/parallel): likewise
  (+description-automatically-generated+): likewise
  (configure-orchestration): new function; construct a dummy project
  and use it to configure orchestration jobs
  (configure-distribution): call `configure-orchestration' instead of
  the removed `configure-jobs'
  (configure-distributions):
  (update-synopsis): mark build-flow-fail commandline option as
  deprecated
  (main): do not process build-flow-fail commandline option; adapted
  changed return values