Added JSON compatibility in src/model/project/json.lisp
The problem is that the multi-line, double-quoted string literals we
used as a JSON extension are processed by folding linebreaks into
spaces according to YAML rules. Thus certain recipes change their
semantics when read as YAML.

To avoid this problem during the transition period, we read recipes as
JSON when possible and switch to YAML otherwise.

* src/model/project/json.lisp: new file; temporarily restored for
  backward compatibility
* src/model/project/package.lisp (header): updated copyright
  (package jenkins.model.project): added exported symbols;
  load-{template,project-spec,distribution}/json-or-yaml
* src/commands/functions-input.lisp (load-project/versioned): changed
  call load-project-spec/{yaml -> json-or-yaml}
* src/commands/command-generate.lisp (generate-load): changed calls
  load-{template,distribution}/{yaml -> json-or-yaml}
* jenkins.project.asd (system jenkins.project): added file
  src/model/project/json.lisp