Use Jenkins API with appropriate content type in lib/jenkins.api/src/api/api.lisp
We previously POSTed using

  text/xml

which I think should work, but is maybe not optimal. The seemingly
correct content type would be

  application/xml

which should defer the encoding decision to the actual content,
i.e. the encoding attribute in the xml processing instruction.

In any case, Jenkins treats both of the above as some non-UTF-8
encoding and jumbles non-ASCII characters accordingly. What works with
Jenkins is

  application/xml; charset=utf-8

which seems wrong, but what do I know.

* lib/jenkins.api/src/api/api.lisp (define-items): use content type
  "application/xml; charset=utf-8"
* changes.sexp (release 0.23): added bugfix entry