Drop version-specific XPath of issues-recorder in lib/jenkins.api/src/api/model/job-publisher.lisp
fixes #51

The version dispatch for the XPath of their minimum-severity slot was
originally required so that the element would be generated with an
acceptable plugin attribute when configuring from scratch. However, as
Robert's experiment indicates, the Warnings NG plugin can now handle
the missing attribute.

Removing the version dispatch eliminates the need for updating the
version list for each plugin release and also simplifies the code.

* lib/jenkins.api/src/api/model/job-publisher.lisp (publisher): for
  the for minimum-severity slot, use an unversioned XPath expression
* changes.sexp (Release 0.34): added bugfix entry

Co-authored-by: Jan Moringen <jmoringe@techfak.uni-bielefeld.de>